Crate graphtest

an examplatory graph module with minimal functionality

Modules

ops
a module for iterating over the indices of collections implementing the std::ops::Index trait

Structs

Node
a node for any data with adjacency information

VGraph
a graph that stores its nodes as a Vec

Traits

SimpleGraph
A simple graph implementation for any collection that implements SuperIndex and IntoIterator.