Module graphtest::ops

a module for iterating over the indices of collections implementing the std::ops::Index trait

With the help of the accepted answer of this stackoverflow question

Cons:

SuperIndex indices cannot be quite as generic as the ones for Index, e.g. slices are not allowed.

Need to explicitly implement SuperIndex for all my types

In each implementation, I must make sure that the two iterators are ordered consistently

Traits

SuperIndex
trait for implementing over the indices of collections that implement std::ops::Index.