Trait graphtest::ops::SuperIndex

pub trait SuperIndex<'a, Idx>: Index<Idx> {
    type IndexIter: Iterator<Item = Idx>;
    type ItemIter: Iterator;
    fn enumerate(&'a self) -> Zip<Self::IndexIter, Self::ItemIter>;
}

trait for implementing over the indices of collections that implement std::ops::Index.

It adds the enumerate function that returns an Enumerate<IndexIter,ItemIter> as an iterator.

Associated Types

type IndexIter: Iterator<Item = Idx>

type ItemIter: Iterator

Loading content...
Required methods

fn enumerate(&'a self) -> Zip<Self::IndexIter, Self::ItemIter>

enumerates over the indices and items of a collection

Loading content...